+++ examples/futures/futures_trading/OrderTrade/OrderTrade.go
@@ -15,6 +15,7 @@ func main() {
 	var yourApiKey = "Your API KEY"
 
 	client := bitmart.NewClient(bitmart.Config{
+		Url:           bitmart.API_URL_V2_PRO,
 		ApiKey:        yourApiKey,
 		TimeoutSecond: 5,
 	})