+++ examples/futures/futures_trading/TradeFeeRate/TradeFeeRate.go
@@ -6,17 +6,23 @@ import (
 )
 
 /*
-GET /contract/public/depth
-Doc: https://developer-pro.bitmart.com/en/futures/#get-market-depth
+GET /contract/private/trade-fee-rate
+Doc: https://developer-pro.bitmart.com/en/futuresv2/#get-trade-fee-rate-keyed
 */
 func main() {
+
+	var yourApiKey = "your api key"
+
 	client := bitmart.NewClient(bitmart.Config{
 		Url:           bitmart.API_URL_V2_PRO,
+		ApiKey:        yourApiKey,
 		TimeoutSecond: 5,
 	})
 
-	// Get Market Depth
-	var ac, err = client.GetContractDepth("BTCUSDT")
+	// Get Trade Fee Rate (KEYED)
+	var ac, err = client.GetContractTradeFeeRate(
+		"BTCUSDT",
+	)
 	if err != nil {
 		log.Panic(err)
 	} else {