+++ examples/futures/market_data/CurrentFundingRate/CurrentFundingRate.go
@@ -10,7 +10,10 @@ GET /contract/public/funding-rate
 Doc: https://developer-pro.bitmart.com/en/futures/#get-current-funding-rate
 */
 func main() {
-	client := bitmart.NewClient(bitmart.Config{TimeoutSecond: 5})
+	client := bitmart.NewClient(bitmart.Config{
+		Url:           bitmart.API_URL_V2_PRO,
+		TimeoutSecond: 5,
+	})
 
 	// Get Current Funding Rate
 	var ac, err = client.GetContractFundingRate("BTCUSDT")