+++ examples/futures/market_data/Kine/Kine.go
@@ -11,7 +11,10 @@ GET /contract/public/kline
 Doc: https://developer-pro.bitmart.com/en/futures/#get-k-line
 */
 func main() {
-	client := bitmart.NewClient(bitmart.Config{TimeoutSecond: 5})
+	client := bitmart.NewClient(bitmart.Config{
+		Url:           bitmart.API_URL_V2_PRO,
+		TimeoutSecond: 5,
+	})
 
 	// Get K-line
 	now := time.Now().Unix()